+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Removed jQuery from the editor controller.
 - Removed jQuery from the updater.
 - Removed jQuery from the tag cloud dialog.
+- Removed jQuery from the stats dialog.
 
 # 1.8.4